feat: introduce an in-memory CourtRepository implementation and register it in the dependency injection container.
