+++ src/infrastructure/container.ts
@@ -2,7 +2,7 @@ import { InMemoryUserRepository } from './repositories/in-memory-user.repository
 import { PaymentRepositoryImpl } from './repositories/payment-repository.impl.js';
 import { BcryptPasswordHasher } from './services/bcrypt-password.hasher.js';
 import { UuidIdGenerator } from './services/uuid-id.generator.js';
-import { InMemoryCourtRepository } from './repositories/in-memory-court.repository.js';
+import { CourtRepositoryImpl } from './repositories/court.repository.impl.js';
 import { InMemorySportRepository } from './repositories/in-memory-sport.repository.js';
 import { UserRepositoryImpl } from './repositories/user.repository.impl.js';
 
@@ -13,5 +13,5 @@ export const userRepository = new UserRepositoryImpl();
 export const paymentRepository = new PaymentRepositoryImpl();
 export const passwordHasher = new BcryptPasswordHasher();
 export const idGenerator = new UuidIdGenerator();
-export const courtRepository = new InMemoryCourtRepository();
+export const courtRepository = new CourtRepositoryImpl();
 export const sportRepository = new InMemorySportRepository();
